requirements.txt
python-telegram-bot
google-generativeai
pymongo
requests
python-dotenv
beautifulsoup4
PyPDF2